app/requirements.txt
fastapi
uvicorn
torch
transformers
streamlit
pillow
diffusers